Move AppShell to container
