+++ src/app.js
@@ -1,158 +1,11 @@
-import ErrorBoundary from './components/ErrorBoundary'
-import Footer from '@twreporter/react-components/lib/footer'
-import Header from '@twreporter/universal-header/dist/containers/header'
-import PropTypes from 'prop-types'
+
+import AppShell from './containers/app-shell'
 import React from 'react'
-import WebPush from './containers/web-push'
 import getRoutes from './routes'
-import memoize from 'memoize-one'
-import mq from '@twreporter/core/lib/utils/media-query'
-import styled from 'styled-components'
-import uiConst from './constants/ui'
-import uiManager from './managers/ui-manager'
-import values from 'lodash/values'
 import { Provider } from 'react-redux'
 import { Switch, Route } from 'react-router-dom'
 import { colors, lineHeight, typography } from './themes/common-variables'
 import { createGlobalStyle } from 'styled-components'
-
-const _ = {
-  values
-}
-
-const AppBox = styled.div`
-  background-color: ${props => props.backgroundColor};
-`
-
-const ContentBlock = styled.div`
-  position: relative;
-`
-
-const TransparentHeader = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  z-index: 1;
-  ${mq.mobileOnly`
-    position: relative;
-  `}
-`
-
-// TODO add `pink` theme to universal-header
-const PinkBackgroundHeader = styled.div`
-  position: relative;
-  background-color: #fabcf0;
-`
-
-class AppShell extends React.PureComponent {
-  static propTypes = {
-    releaseBranch: PropTypes.string.isRequired,
-    layoutObj: PropTypes.shape({
-      header: PropTypes.oneOf(
-        _.values(uiConst.header)
-      ),
-      footer: PropTypes.oneOf(
-        _.values(uiConst.footer)
-      ),
-      backgroundColor: PropTypes.string
-    })
-  }
-
-  static defaultProps = {
-    layoutObj: {
-      header: uiConst.header.default,
-      footer: uiConst.footer.default,
-      backgroundColor: '#f1f1f1'
-    }
-  }
-
-  renderHeader() {
-    const { layoutObj, releaseBranch } = this.props
-
-    switch(layoutObj.header) {
-      case uiConst.header.none: {
-        return null
-      }
-      case uiConst.header.pink: {
-        return (
-          <PinkBackgroundHeader>
-            <Header
-              theme="transparent"
-              releaseBranch={releaseBranch}
-              isLinkExternal={false}
-            />
-          </PinkBackgroundHeader>
-        )
-      }
-      case uiConst.header.photo: {
-        return (
-          <Header
-            theme="photography"
-            releaseBranch={releaseBranch}
-            isLinkExternal={false}
-          />
-        )
-      }
-      case uiConst.header.transparent: {
-        return (
-          <TransparentHeader>
-            <Header
-              theme="transparent"
-              releaseBranch={releaseBranch}
-              isLinkExternal={false}
-            />
-          </TransparentHeader>
-        )
-      }
-      default: {
-        return (
-          <Header
-            theme="normal"
-            releaseBranch={releaseBranch}
-            isLinkExternal={false}
-          />
-        )
-      }
-    }
-  }
-
-  renderFooter() {
-    const { layoutObj } = this.props
-
-    switch(layoutObj.footer) {
-      case uiConst.footer.none: {
-        return null
-      }
-      case uiConst.footer.default:
-      default: {
-        return <Footer />
-      }
-    }
-  }
-
-  render() {
-    const {
-      layoutObj,
-      children
-    } = this.props
-
-    return (
-      <ErrorBoundary>
-        <AppBox
-          backgroundColor={layoutObj.backgroundColor}
-        >
-          <WebPush />
-          <ContentBlock>
-            {this.renderHeader()}
-            {children}
-            {this.renderFooter()}
-          </ContentBlock>
-        </AppBox>
-      </ErrorBoundary>
-    )
-  }
-}
 
 const GlobalStyle = createGlobalStyle`
   html {
@@ -249,12 +102,9 @@ const GlobalStyle = createGlobalStyle`
   }
 `
 
+
 export default class App extends React.Component {
-  buildLayoutObjForAppShell = memoize((header, footer, backgroundColor) => ({
-    header,
-    footer,
-    backgroundColor
-  }))
+
 
   render() {
     const routes = getRoutes()
@@ -263,11 +113,9 @@ export default class App extends React.Component {
     return (
       <Provider store={reduxStore}>
         <Route render={props => {
-          const reduxState = reduxStore.getState()
-          const { header, footer, backgroundColor } = uiManager.getLayoutObj(reduxState, props.location)
           return (
             <AppShell
-              layoutObj={this.buildLayoutObjForAppShell(header, footer, backgroundColor)}
+              location={props.location}
               releaseBranch={releaseBranch}
             >
               <Switch>
